move to spe subconfig
